update mileage integrated

+++ app/src/pitstop/java/com/pitstop/ui/dialog/MileageDialogPresenter.kt
@@ -35,7 +35,7 @@ class MileageDialogPresenter(private val usecaseComponent: UseCaseComponent) {
             override fun onCarRetrieved(car: Car?, dealership: Dealership?, isLocal: Boolean) {
                 if (car != null){
                     view?.showMileage(car.totalMileage.toInt())
-                    view?.setEditText(car.totalMileage.toString())
+                    view?.setEditText(car.totalMileage.toInt().toString())
                 }
             }
